More analytics, crash fix

+++ mobile/src/main/java/rhedox/gesahuvertretungsplan/util/BootReceiver.java
@@ -24,7 +24,7 @@ public class BootReceiver extends BroadcastReceiver {
 
             SharedPreferences prefs = PreferenceManager.getDefaultSharedPreferences(context);
             String mode = prefs.getString(PreferenceFragment.PREF_NOTIFICATION_MODE, null);
-            LocalTime time = LocalTime.fromMillisOfDay(prefs.getInt(PreferenceFragment.PREF_NOTIFICATION_TIME, 0));
+            LocalTime time = LocalTime.fromMillisOfDay(prefs.getLong(PreferenceFragment.PREF_NOTIFICATION_TIME, 0));
 
             if(!"none".equals(mode)) {
                 @AlarmReceiver.NotificationFrequency int notificationFrequency;
